ResourceMonitor v2.1.0 / PerfDatOSProcessor added, Windows7 process exit problem solved. Readme v 1.1.0 updated

+++ ResourceMonitor2008_2/ResourceMonitor2008_2/PerfData.h
@@ -59,7 +59,7 @@ protected:
 	vector<CString>			propertyNames;
 
 	VARIANT					propertyVal;
-	ULONGLONG				ID;
+	int						ID;
 
 	virtual void SetDataObj(int index) abstract;
 	virtual void SetTableInstance() abstract;
